Food products stay liked if they have been liked by a user

+++ src/main/java/com/project/final_year_project/model/java/service/UserFoodProductRatingService.java
@@ -43,9 +43,14 @@ public class UserFoodProductRatingService {
         return userFoodProductRatingRepository.findById(userFoodProductRatingID);
     }
 
-    public void removeRating(Long userID, String code, String relationship) {
+    public boolean removeRating(Long userID, String code, String relationship) {
         UserFoodProductRatingID ratingID = new UserFoodProductRatingID(userID, code, getRatingType(relationship));
-        userFoodProductRatingRepository.deleteById(ratingID);
+        if (userFoodProductRatingRepository.existsById(ratingID)) {
+            userFoodProductRatingRepository.deleteById(ratingID);
+            System.out.println("Rating has been removed");
+            return true;
+        }
+        return false;
     }
 
     public Rating getRatingType(String relationship) {
